[CONFIG] Configure SonarCloud indexed files (attempt 4)
[CONFIG] Configure SonarCloud indexed files (attempt 5)

[CONFIG] Configure SonarCloud indexed files (attempt 6)

[CONFIG] Configure SonarCloud indexed files (attempt 7)

[CONFIG] Configure SonarCloud indexed files (idk anymore)

[CONFIG] Configure SonarCloud indexed files (idk anymore asdaf)

[CONFIG] Configure SonarCloud indexed files (idk anymore asdaf2)

[CONFIG] Configure SonarCloud indexed files (asfagadf)

[CONFIG] Configure SonarCloud indexed files (n)

[CONFIG] Configure SonarCloud indexed files (na)

[CONFIG] Configure SonarCloud indexed files (nasda)

[CONFIG] Configure SonarCloud indexed files (nasfafgsda)

[CONFIG] Configure SonarCloud indexed files (nagfsh)

[CONFIG] Configure SonarCloud indexed what

[CONFIG] Configure SonarCloud indexed when

[CONFIG] Configure SonarCloud indexed where

Configuring SonarCloud and rename package
