+++ .github/workflows/sonarcloud.yml
@@ -1,26 +1,21 @@
 name: SonarCloud Analysis
 on:
-  # Trigger analysis when pushing in master or pull requests, and when creating
-  # a pull request.
   push:
   pull_request:
 
 jobs:
   build:
     name: Build
-    runs-on: ubuntu-22.04
+    runs-on: ubuntu-latest
     steps:
-      - name: Check out the Git repository
-        uses: actions/checkout@v4
+      - uses: actions/checkout@v4
         with:
-          # Shallow clones should be disabled for a better relevancy of analysis
-          fetch-depth: 0
-      - name: Set up Java Toolchain
+          fetch-depth: 0  # Shallow clones should be disabled for a better relevancy of analysis
+      - name: Set up JDK 21
         uses: actions/setup-java@v4
         with:
-          distribution: "temurin"
-          java-version: "21"
-          cache: "gradle"
+          java-version: 21
+          distribution: 'temurin' # Alternative distribution options are available
       - name: Cache SonarCloud packages
         uses: actions/cache@v4
         with:
@@ -36,18 +31,14 @@ jobs:
           key: ${{ runner.os }}-gradle-${{ hashFiles('**/*.gradle') }}
           restore-keys: ${{ runner.os }}-gradle
       - name: Build and analyze
-        run: ./gradlew build sonar --info
         env:
-          # Needed to get some information about the pull request, if any
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          # SonarCloud access token should be generated from https://sonarcloud.io/account/security/
           SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
           PRODUCTION: test
+        run: ./gradlew build sonar --info
       - name: Run tests with coverage
         run: ./gradlew test jacocoTestReport
-        env:
-          PRODUCTION: test
       - name: Run analysis
         env:
-            SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
-        run: ./gradlew sonar -Dsonar.projectKey=B05-Advpro_hoomgroom-be-auth -Dsonar.organization=b05-advpro -Dsonar.host.url=https://sonarcloud.io -Dsonar.login=${{ secrets.SONAR_TOKEN }} -Dsonar.coverage.jacoco.xmlReportPaths=build/reports/jacoco/test/jacocoTestReport.xml
+          SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
+        run: ./gradlew sonar -Dsonar.projectKey=B05-Advpro_hoomgroom-be-auth -Dsonar.organization=b05-advpro -Dsonar.host.url=https://sonarcloud.io -Dsonar.token=${{ secrets.SONAR_TOKEN }} -Dsonar.coverage.jacoco.xmlReportPaths=build/reports/jacoco/test/jacocoTestReport.xml -Dsonar.junit.reportPaths=build/test-results/test
