+++ sonar-project.properties
@@ -1,0 +1,6 @@
+sonar.organization=b05-advpro
+sonar.projectKey=B05-Advpro_hoomgroom-be-auth
+
+# relative paths to source directories. More details and properties are described
+# in https://sonarcloud.io/documentation/project-administration/narrowing-the-focus/
+sonar.sources=.
