+++ src/main/java/id/ac/ui/cs/advprog/hoomgroomauth/dto/AuthenticationRequest.java
@@ -1,0 +1,15 @@
+package id.ac.ui.cs.advprog.hoomgroomauth.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class AuthenticationRequest {
+    private String username;
+    private String password;
+}
