+++ src/test/java/id/ac/ui/cs/advprog/hoomgroomauth/HoomgroomBeAuthApplicationTests.java
@@ -1,13 +1,9 @@
-package id.ac.ui.cs.advprog.hoomgroom.auth;
+package id.ac.ui.cs.advprog.hoomgroomauth;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
 class HoomgroomBeAuthApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
 
 }
